requirements.txt
aniso8601==8.0.0
bcrypt==3.1.7
cffi==1.13.2
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
papi==1.0
paramiko==2.7.1
pkg-resources==0.0.0
pycparser==2.19
PyJWT==1.7.1
PyNaCl==1.3.0
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==1.0.0rc1
